Moved signum header

+++ include/sys/signal_num.h
@@ -51,34 +51,34 @@
  *  S   Stop the process.
  *  C   Continue the process, if it is stopped; otherwise, ignore the signal.
  */
-#define SIGHUP  ((int)1)    /*!< (T) Hangup. */
-#define SIGINT  ((int)2)    /*!< (T) Terminal interrupt signal. */
-#define SIGQUIT ((int)3)    /*!< (A) Terminal quit signal. */
-#define SIGILL  ((int)4)    /*!< (A) Illegal instruction. */
-#define SIGTRAP ((int)5)    /*!< (A) Trace/breakpoint trap. */
-#define SIGABRT ((int)6)    /*!< (A) Process abort signal. */
-#define SIGCHLD ((int)7)    /*!< (I) Child process terminated, stopped. */
-#define SIGFPE  ((int)8)    /*!< (A) Erroneous arithmetic operation. */
-#define SIGKILL ((int)9)    /*!< (T) Kill (cannot be caught or ignored). */
-#define SIGBUS  ((int)10)   /*!< (A) Access to an undefined portion of a memory
-                             *   object. */
-#define SIGSEGV ((int)11)   /*!< (A) Invalid memory reference. */
-#define SIGCONT ((int)12)   /*!< (C) Continue executing, if stopped. */
-#define SIGPIPE ((int)13)   /*!< (T) Write on a pipe with no one to read it. */
-#define SIGALRM ((int)14)   /*!< (T) Alarm clock. */
-#define SIGTERM ((int)15)   /*!< (T) Termination signal. */
-#define SIGSTOP ((int)16)   /*!< (S) Stop executing (cannot be caught or
-                             *   ignored). */
-#define SIGTSTP ((int)17)   /*!< (S) Terminal stop signal. */
-#define SIGTTIN ((int)18)   /*!< (S) Background process attempting read. */
-#define SIGTTOU ((int)19)   /*!< (S) Background process attempting write. */
-#define SIGUSR1 ((int)20)   /*!< (T) User-defined signal 1. */
-#define SIGUSR2 ((int)21)   /*!< (T) User-defined signal 2. */
-#define SIGSYS  ((int)22)   /*!< (A) Bad system call. */
-#define SIGURG  ((int)23)   /*!< (I) High bandwidth data is available at a
-                             *   socket. */
-#define SIGINFO ((int)24)   /*!< (I) Info request. */
-#define SIGPWR  ((int)25)   /*!< (T) System power failure. */
+#define SIGHUP  1   /*!< (T) Hangup. */
+#define SIGINT  2   /*!< (T) Terminal interrupt signal. */
+#define SIGQUIT 3   /*!< (A) Terminal quit signal. */
+#define SIGILL  4   /*!< (A) Illegal instruction. */
+#define SIGTRAP 5   /*!< (A) Trace/breakpoint trap. */
+#define SIGABRT 6   /*!< (A) Process abort signal. */
+#define SIGCHLD 7   /*!< (I) Child process terminated, stopped. */
+#define SIGFPE  8   /*!< (A) Erroneous arithmetic operation. */
+#define SIGKILL 9   /*!< (T) Kill (cannot be caught or ignored). */
+#define SIGBUS  10  /*!< (A) Access to an undefined portion of a memory
+                     *   object. */
+#define SIGSEGV 11  /*!< (A) Invalid memory reference. */
+#define SIGCONT 12  /*!< (C) Continue executing, if stopped. */
+#define SIGPIPE 13  /*!< (T) Write on a pipe with no one to read it. */
+#define SIGALRM 14  /*!< (T) Alarm clock. */
+#define SIGTERM 15  /*!< (T) Termination signal. */
+#define SIGSTOP 16  /*!< (S) Stop executing (cannot be caught or
+                     *   ignored). */
+#define SIGTSTP 17  /*!< (S) Terminal stop signal. */
+#define SIGTTIN 18  /*!< (S) Background process attempting read. */
+#define SIGTTOU 19  /*!< (S) Background process attempting write. */
+#define SIGUSR1 20  /*!< (T) User-defined signal 1. */
+#define SIGUSR2 21  /*!< (T) User-defined signal 2. */
+#define SIGSYS  22  /*!< (A) Bad system call. */
+#define SIGURG  23  /*!< (I) High bandwidth data is available at a
+                     *   socket. */
+#define SIGINFO 24  /*!< (I) Info request. */
+#define SIGPWR  25  /*!< (T) System power failure. */
 
 #endif /* SIGNAL_NUM_H */
